+++ src/utils/vec3.c
@@ -19,7 +19,7 @@ Vec3 vec3_sub(Vec3 a, Vec3 b) {
     return a;
 }
 
-double vec3_dot(Vec3 a, Vec3 b) {
+float vec3_dot(Vec3 a, Vec3 b) {
     ASSERT_VECTOR_VALUE_NOT_NAN(a);
     ASSERT_VECTOR_VALUE_NOT_NAN(b);
     return a.x * b.x + a.y * b.y + a.z * b.z;
@@ -36,7 +36,7 @@ Vec3 vec3_cross(Vec3 a, Vec3 b) {
     return result;
 }
 
-Vec3 vec3_clamp(Vec3 a, double min, double max) {
+Vec3 vec3_clamp(Vec3 a, float min, float max) {
     ASSERT_VECTOR_VALUE_NOT_NAN(a);
     a.r = math_clamp(a.r, min, max);
     a.g = math_clamp(a.g, min, max);
@@ -57,7 +57,7 @@ Vec3 vec3_hadamard(Vec3 a, Vec3 b) {
 
 Vec3 vec3_norm(Vec3 a) {
     ASSERT_VECTOR_VALUE_NOT_NAN(a);
-    double length = vec3_length(a);
+    float length = vec3_length(a);
     if (length != 0) {
         a.x /= length;
         a.y /= length;
@@ -66,12 +66,12 @@ Vec3 vec3_norm(Vec3 a) {
     return a;
 }
 
-double vec3_length(Vec3 a) {
+float vec3_length(Vec3 a) {
     ASSERT_VECTOR_VALUE_NOT_NAN(a);
-    return sqrt(a.x * a.x + a.y * a.y + a.z * a.z);
+    return sqrtf(a.x * a.x + a.y * a.y + a.z * a.z);
 }
 
-Vec3 vec3_offset(Vec3 a, double offset) {
+Vec3 vec3_offset(Vec3 a, float offset) {
     ASSERT_VECTOR_VALUE_NOT_NAN(a);
     assert(!isnan(offset));
     a.x += offset;
@@ -80,7 +80,7 @@ Vec3 vec3_offset(Vec3 a, double offset) {
     return a;
 }
 
-Vec3 vec3_mul(Vec3 a, double b) {
+Vec3 vec3_mul(Vec3 a, float b) {
     ASSERT_VECTOR_VALUE_NOT_NAN(a);
     assert(!isnan(b));
     a.x *= b;
@@ -89,7 +89,7 @@ Vec3 vec3_mul(Vec3 a, double b) {
     return a;
 }
 
-Vec3 vec3_div(Vec3 a, double b) {
+Vec3 vec3_div(Vec3 a, float b) {
     ASSERT_VECTOR_VALUE_NOT_NAN(a);
     assert(!isnan(b));
     assert(b != 0);
